Moved music butto to options screen and fixed how we change screens in title screen and setup screen
